Update for 25.03-BETA
Updated schema and functions that are introduced in 25.03. The code is
still in development, and this drop is intended to allow testing against
25.03-BETA until RELEASE is complete.

The OpenAPI definition now introduces "required" tags to list fields
that are mandatory. They are incomplete, as we go through the list of
parameters in the API to ensure that anything optional (or variably
dependent on something else) is not marked as required. Documentation is
still terse; will be updated accordingly.